Lots of stuff
Rewrote main to have a debug mode (since now assignments dont need to output stuff to terminal). Removed the CHECK_ARGS thing so if you run the program with no input it will automatically run on the program selected on the selected input file(its obvious where it is). if there is one input file passed then the program now goes to the interpreter and exits via that with the approprate exit code. if you run it with the file and -d extension. Debug mode will print all past assignment material. If you are running it on clion it will have the option to enable or disable debug mode in the section labeled for you. Then I made an interpreter skeleton with nothing in it yet. should be a good framework to build on. also updated the makefile.