Little bit of cleanup and program exits differantly
I changed it so that the program finishes inside of Begin. That might need to change but for now if we get back into main after calling Begin there is an issue ISSUE becuase we have checks for the ast is complete and if the program finishes. Anyway hence the return code of 1 in main.